Bump version of Faraday 0.9.0.
This is a dependency update which will bring several projects to the
latest version of Faraday. I have put in pull-requests to the
berkshelf and ridley gems for this as well.

Here's to hoping that no_proxy (environment) support will be merged into
Faraday 0.9.0 branch soon so that all of us poor schmucks in the
enterprise can rejoice.