Completely move to Supermarket
This introduces some breaking changes. First, the defaults are all now
pointing at Supermarket. Additionally, the Supermarket worker now just
consumes the +universe.json+ endpoint instead of downloading thousands of
cookbooks for indexing.

We are currently seeing exceptionally high loads from users indexing the new
Supermarket site, so this would be a great patch to merge. It reduces
bandwidth and prevents the need to download every cookbook just to parse the
metadata.